server/requirements.txt
flask==3.1.*
pyjwt==2.10.*
